move suggestions to parserlibPowered
